+++ src/index.js
@@ -1,14 +1,13 @@
 import './style.scss';
-import { DomModule } from './helpers/domManipulation';
 import { ConfigurePage } from './helpers/configurePage';
 import { projectModule } from './project/projectModule';
 
 const keyStorage = 'todoList';
 
-window.onload = function () {
+window.onload = function initialActions() {
   ConfigurePage.InitializePage(keyStorage);
 };
 
-window.addEventListener('beforeunload', (event) => {
-  if (projectModule.getProjectAmount != 0) projectModule.saveToStorage(keyStorage);
+window.addEventListener('beforeunload', () => {
+  if (projectModule.getProjectAmount !== 0) projectModule.saveToStorage(keyStorage);
 });
