+++ src/task/taskModule.js
@@ -2,17 +2,18 @@ import { taskFactory } from './taskFactory';
 
 export const taskModule = (() => {
   const createTask = () => {
+    const taskArray = [];
     let newIndex = 0;
-    if (taskArray.length != 0) {
+    if (taskArray.length !== 0) {
       newIndex = taskArray[taskArray.length - 1].getIndex() + 1;
     }
     const newTask = taskFactory(newIndex);
     taskArray.push(newTask);
   };
 
-  const getTaskArray = () => taskArray;
   return {
     createTask,
-    getTaskArray,
   };
 })();
+
+export default taskModule;
